refactor: extract matchStringValue
