+++ src/helpers/matchers/matchStringValue.ts
@@ -1,0 +1,16 @@
+import { TextMatch } from '../../matches';
+
+export function matchStringValue(
+  prop: string | undefined,
+  matcher: TextMatch
+): boolean {
+  if (!prop) {
+    return false;
+  }
+
+  if (typeof matcher === 'string') {
+    return prop === matcher;
+  }
+
+  return Boolean(prop.match(matcher));
+}
